pyproject.toml
[project]
name = "imagine"
version = "0.1.0"
description = "Imagine Server"
authors = [{ name = "Christian Bick", email = "christian.bick@bitsuppliers.com" }]
requires-python = "~=3.12"
dependencies = [
    "owlready2>=0.46,<0.47",
    "flask~=3.0",
    "python-dotenv>=1.0.1,<2",
    "flask-cors>=5.0.0,<6",
    "waitress>=3.0.1,<4",
    "google-genai>=1.27.0",
    "google-cloud-storage>=2.19.0",
    "numpy>=2.3.0",
    "google-cloud-aiplatform>=1.105.0",
    "pymupdf>=1.26.0",
    "setuptools>=80.9.0",
]

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest~=8.3",
    "assertpy~=1.1",
]

